Implement bruteforce game solver
This patch creates a new toolbar button a user can click to have
Aisleriot attempt to solve the game at hand, optionally continuing until
a winning game is found. This patch written specifically for Auld Lang
Syne and it's expected this is the only game it works for thus far, but
it's intended that more games will be functional in the future.