Fix infinite loop bug
The primary goal of this commit is to fix a pretty common scenario in
which the backtracking algo would get stuck and play the same moves over
and over. The choice increment logic fails to account for the case of a
card being playable onto more than one foundation, so before adding a
playable slot into available_choices this patch first checks whether the
slot is already in the array.

The fix is only one or two lines of this commit, the rest is additional
fluff:
 * Add a cool looking (but mostly useless) debug aid
 * Add an option for multi-game solution search
 * Allow solution search to be interrupted
 * Quit on main window close